requirements.txt
opencv-python==4.10.0.84
python-docx==1.1.2
matplotlib==3.8.2
pdf2image==1.17.0
tqdm==4.66.5
unoserver==2.1
unotools==0.3.3
beautifulsoup4==4.12.3
requests==2.32.3
# pillow==10.4.0
pillow-simd==9.5.0.post2
augraphy==8.2.6